Create class (LocationService) that get the users current lat long coordinates. And also a another clas (HttpTravelInfoRequest) that makes a Get request to info of buses based on user lat long

+++ Entities/TravelInfomation.swift
@@ -13,9 +13,10 @@ struct TravellInfomation: Codable {
 }
 
 struct BusInfo: Codable {
+    
     var naptanId: String
     var commonName: String
-    var distance: Int
+    var distance: Double
     var additionalProperties: [AdditionalProperties]
 }
